% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse_crontab.R
\name{cleanse_crontab}
\alias{cleanse_crontab}
\title{Cleanse a crontab file of the DayNotis entries}
\usage{
cleanse_crontab(file = "crontab")
}
\arguments{
\item{file}{file name (character/scalar)}

\item{new_file}{name of new crontab}
}
\description{
Cleanse a crontab file of the DayNotis entries
}
