% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_crontab.R
\name{make_new_crontab}
\alias{make_new_crontab}
\title{Make a new crontab by combining current crontab with DayNotis crontab}
\usage{
make_new_crontab(file1 = "crontab", file2 = "DayNotis_crontab")
}
\arguments{
\item{file1}{current crontab}

\item{file2}{DayNotis crontab}
}
\description{
Make a new crontab by combining current crontab with DayNotis crontab
}
