% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_crontabs.R
\name{merge_crontabs}
\alias{merge_crontabs}
\title{Make a new crontab by combining current crontab with DayNotis crontab}
\usage{
merge_crontabs(file1 = "crontab", file2 = "crontab_DayNotis",
  outfile = "new_crontab")
}
\arguments{
\item{file1}{current crontab}

\item{file2}{DayNotis crontab}

\item{outfile}{resulting crontab}
}
\description{
Make a new crontab by combining current crontab with DayNotis crontab
}
