% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification_sound.R
\name{notification_sound}
\alias{notification_sound}
\title{Return the MacOSX notificaiton sound for the given the priority level}
\usage{
notification_sound(level)
}
\arguments{
\item{level}{priority (character/scalar)}
}
\description{
Return the MacOSX notificaiton sound for the given the priority level
}
