% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_notification_cron.R
\name{write_notification_cron}
\alias{write_notification_cron}
\title{Write a single crontab line for creating a notification}
\usage{
write_notification_cron(time, event)
}
\arguments{
\item{time}{time of day (HHMM)}

\item{event}{event details (body, title, level)}
}
\description{
Write a single crontab line for creating a notification
}
