% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeUtils.R
\name{write_parameter}
\alias{write_parameter}
\title{Convert a yml parameter to text}
\usage{
write_parameter(x, pre_text = NULL)
}
\arguments{
\item{x}{parameter text}

\item{pre_text}{pre-text}
}
\description{
Convert a yml parameter to text
}
